Wrote reset_password function
It's 90% done, not sure how to update the user's password on the databse when filtering by their e-mail with the way the SQL stuff is setup.

Also, I tried testing it but I didn't get any e-mail when I did. I assume I've done it wrong in login.php, not sure what to do there exactly.